Add files via upload
added authentication
completed currencies page
completed signup & login page

+++ src/controllers/currencyControllers.js
@@ -13,49 +13,48 @@ controller.use(express.json());
 // Create a new currency
 export const createCurrency = async (req, res) => {
     try {
-        const currency = new Currency(req.body);
-        await currency.save();
-        res.status(201).send(currency);
+        if (req.isAuthenticated()) {
+            const newcurrency = new Currency(req.body);
+            await newcurrency.save();
+            res.status(201).redirect("/currencies");
+        } else {
+            res.redirect("/login");
+        }
     } catch (error) {
-        res.status(400).send(error);
+        res.status(400).send(error.message);
     }
 };
+
+
 
 // Get all currencies
 export const getAllCurrencies = async (req, res) => {
     try {
-        const currencies = await Currency.find();
-        const data = currencies.map(currency => {
-            return {
-                _id: currency._id,
-                icon: currency.icon,
-                nameInArabic: currency.nameInArabic,
-                code: currency.code,
-                exchRate: currency.exchRate,
-                priorityCu: currency.priorityCu
-            }
-        })
-
-        console.log(data)
-        res.status(200).send(data);
+        if (req.isAuthenticated()) {
+            const currencies = await Currency.find().sort({priorityCu: 1});
+            res.status(200).render("currency-management.ejs", {currencies: currencies})
+        } else {
+            res.redirect("/login");
+        }
     } catch (error) {
-        res.status(500).send(error);
+        res.status(500).send(error.message);
     }
 };
 
 // Get a currency by ID
 export const getCurrencyById = async (req, res) => {
     try {
-        
-        console.log("\nreq params:", req.params);
-        const currency = await Currency.findById(req.params.id);
-        console.log("\nquery result:", currency);
-        
-        res.status(200).json(currency);
+        if (req.isAuthenticated()) {
+            const currency = await Currency.findById(req.params.id);   
+            console.log(currency);         
+            res.status(200).json(currency);
+        } else {
+            res.redirect("/login")
+        }
 
     } catch (error) {
         console.log("Error:", error.message);
-        res.status(500).json("error fetching currency");
+        res.status(500).json(error.message);
     }
 };
 
@@ -63,32 +62,38 @@ export const getCurrencyById = async (req, res) => {
 // Patch for currency all fields required
 export const patchCurrencyById = async (req, res) => {
     try {
-
-        const updates = req.body;
-        console.log(updates);
-        
-        const currency = await Currency.findByIdAndUpdate(req.params.id, updates, { new: true, runValidators: true });
-        
-
-        res.status(200).json(currency);
+        console.log("patch currency");
+        if (req.isAuthenticated()) {
+            const updates = req.body;
+            console.log(updates);
+            await Currency.findByIdAndUpdate(req.params.id , updates, { new: true, runValidators: true });
+            res.status(200).json({message: "patch successeful"})
+        } else {
+            res.redirect("/login");
+        }
     } catch (error) {
-        res.status(400).json(error.message);
+        res.status(400).send(error.message);
     }
 };
 
 // Delete a currency by ID
 export const deleteCurrencyById = async (req, res) => {
     try {
-        console.log(req.params);
-        const result = await Currency.findByIdAndDelete(req.params.id);
-        console.log(result);
-
-
-        res.status(200).redirect("/currencies");
+        if (req.isAuthenticated()) {
+            console.log(req.params);
+            await Currency.findByIdAndDelete(req.params.id);
+            res.status(200).json({message: "delete successful"});
+        } else {
+            res.redirect("/login");
+        }
     } catch (error) {
         res.status(500).json(error.message);
     }
 };
+
+
+
+
 
 export default {
     createCurrency,
